Can't call git commands on Heroku.
After you push to Heroku, Heroku removes the `.git` folder, so any `git` commands will result in `fatal: Not a git repository (or any of the parent directories): .git`.

The solution is to set the environment variable on Heroku with a git pre-push hook as explained here:
http://stackoverflow.com/a/22702304

The purpose of this code is to allow the browser cache to be invalidated when a new release is pushed to Heroku by setting the `Etag` to the latest commit hash.