Replace Memcache with Redis
Why:

The new readthis gem, which works with Redis, is about 8x faster than dalli with memcached based on my tests with ApacheBench. Furthermore, with my fork of redis-rack-cache, we can specify when entries expire. Before, the cache would keep filling up and entries would never expire, which I'm pretty sure was the cause of the R14 memory errors on Heroku.