Remove Google translation features
Why: The Google translation widget is no longer supported by Google,
and the only way I was able to to make it work was to allow unsafe
inline scripts in the CSP (Content Security Policy) settings, which is
not recommended.

Allowing unsafe inline scripts means any website can inject scripts,
which defeats the purpose of having a CSP config that allows you to
choose which domains to allow.