Update Ruby to 2.7.4 and Node to 16.13.0
On my M1 Mac, in order for "bin/setup" to complete successfully, this required updating ffi, puma, and nokogiri at the same time, and webpacker also needed to be updated.

When bin/setup ran successfully, I started the server with "rails s", but when I went to localhost:3000, I got "no site key specified", which seems to come from the recaptcha gem. I had to set the ENV vars locally.