Drop support for IE 8 and 9 and remove unused code
The CodeQL security analysis was flagging code in `util.js` that turned
out to be  unused. It was related to the Google translate feature that
was removed a while back.

CodeQL was also flagging `selectivizr.js`, which is only used to support
IE 8 and 9. However, those browsers are not supported anymore by
Microsoft, and recent visitor analytics for smc-connect.org show that
no one is using IE at all.

Google Maps also does not support those old versions of IE, and displays
a warning that recommends using a different browser.

All of the files in the `vendor` folder were only used to support IE 8
and 9, so we can remove them all.