Send the emails from a verified email
SendGrid requires the email sender to be verified. Previously we were setting the sender to the email address of the person submitting the form, and if they didn't enter an email address (it's optional), then we would set the sender to "anonymous@none.com", which SendGrid won't allow.

So, instead, we set the sender to a verified email.