Show Google logo on maps
I got an email from the Google Maps API Compliance team saying smc-connect.org was non-compliant because the maps aren't displaying the logo.

After some debugging, it turns out we weren't doing it on purpose. There was a CSS styling applied to the "main" class that made that part of the map disappear.

Disabling that position setting then caused the filter checkboxes to disappear. So I moved that setting to just the "toggle-group" class and now everything looks great.