Update Ruby to 3.1.2
The latest Heroku stack (22) only supports Ruby 3.1.2
To be able to bundle install, I had to update both derailed_benchmarks and listen at the same time because heapy and ruby_dep depended on Ruby 2.x before.

But then running `script/test` threw errors from compass via compass-rails due to SortedSet.

Adding the sorted_set gem fixed that.

Then got error about `net/smtp` missing. That's because that gem was removed from Ruby (similar to webrick).

The fix was to add the latest RC of the `mail` gem.

Then I got errors from kaminari about wrong number of arguments for paginate. Updating kaminari to 1.2.2 fixed it but we're back to the bug with the total count in `page_entries_info`.

So I forked kaminari and fixed it, and now it's pointing to my branch until kaminari merges it.